certificate-sign-data/requirements.txt
azure.identity==1.3.1
azure.keyvault==4.1.0